+++ src/naming/redo/automatic_request/persistent_instance_request.rs
@@ -1,13 +1,13 @@
+use async_trait::async_trait;
 use std::sync::Arc;
-use tonic::async_trait;
+use tracing::Instrument;
 use tracing::debug;
 use tracing::error;
 use tracing::instrument;
-use tracing::Instrument;
 
 use crate::common::remote::generate_request_id;
-use crate::common::remote::grpc::message::GrpcMessageData;
 use crate::common::remote::grpc::NacosGrpcClient;
+use crate::common::remote::grpc::message::GrpcMessageData;
 use crate::naming::message::request::PersistentInstanceRequest;
 use crate::naming::message::response::InstanceResponse;
 use crate::naming::redo::AutomaticRequest;